Static bad.

+++ Content.Shared/Beam/Components/SharedBeamComponent.cs
@@ -50,7 +50,7 @@ public abstract partial class SharedBeamComponent : Component
     /// Beams of the same family have unique index. Used to make sure lightning hits an entity no more than once.
     /// </summary>
     [ViewVariables(VVAccess.ReadOnly)]
-    public ushort BeamIndex;
+    public uint BeamIndex;
 }
 
 /// <summary>
